+++ src/pages/index.tsx
@@ -10,7 +10,7 @@ import PayFailPage from "./PayFail/PayFailPage";
 import Mypage from "./Mypage/Mypage";
 import ProfilePage from "./Mypage/ProfilePage/ProfilePage";
 import PostManagePage from "./Mypage/admin/PostManage/PostManagePage";
-import PostCRUDPage from "./Mypage/admin/PostCRUD/PostCRUDPage";
+import PostUploadPage from "./Mypage/admin/PostUpload/PostUploadPage";
 import PostUpdatePage from "./Mypage/admin/PostUpdate/PostUpdatePage";
 import AdminReservationManagePage from "./Mypage/admin/ReservationManage/ReservationManagePage";
 import LikeManagePage from "./Mypage/user/LikeManage/LikeManagePage";
@@ -31,7 +31,7 @@ export {
   Mypage,
   ProfilePage,
   PostManagePage,
-  PostCRUDPage,
+  PostUploadPage,
   PostUpdatePage,
   AdminReservationManagePage,
   UserReservationManagePage,
